change id type from Integer to Long

+++ src/main/java/cz/patyk/invoicesystem_be/repositories/TariffRepository.java
@@ -3,5 +3,5 @@ package cz.patyk.invoicesystem_be.repositories;
 import cz.patyk.invoicesystem_be.entities.Tariff;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TariffRepository extends JpaRepository<Tariff, Integer> {
+public interface TariffRepository extends JpaRepository<Tariff, Long> {
 }
